- Fix warnings - Rename DYNAMIC to RESIZABLE_DYNAMIC

+++ rawrbox.render/include/rawrbox/render/textures/pack.hpp
@@ -6,21 +6,21 @@
 
 namespace rawrbox {
 	struct PackNode {
-		uint16_t x = 0;
-		uint16_t y = 0;
-		uint16_t width = 0;
-		uint16_t height = 0;
+		uint32_t x = 0;
+		uint32_t y = 0;
+		uint32_t width = 0;
+		uint32_t height = 0;
 
 		std::vector<uint8_t> data = {};
 
 		std::unique_ptr<PackNode> left = nullptr;
 		std::unique_ptr<PackNode> right = nullptr;
 
-		bool canInsertNode(uint16_t insertedWidth, uint16_t insertedHeight);
-		std::optional<std::reference_wrapper<rawrbox::PackNode>> InsertNode(uint16_t width, uint16_t height);
+		bool canInsertNode(uint32_t insertedWidth, uint32_t insertedHeight);
+		std::optional<std::reference_wrapper<rawrbox::PackNode>> InsertNode(uint32_t width, uint32_t height);
 
 		PackNode() = default;
-		PackNode(uint16_t _x, uint16_t _y, uint16_t _w, uint16_t _h) : x(_x), y(_y), width(_w), height(_h){};
+		PackNode(uint32_t _x, uint32_t _y, uint32_t _w, uint32_t _h) : x(_x), y(_y), width(_w), height(_h){};
 	};
 
 	class TexturePack : public rawrbox::TextureBase {
@@ -30,7 +30,7 @@ namespace rawrbox {
 		bool _pendingUpdate = false;
 
 	public:
-		explicit TexturePack(uint16_t size = 1024U);
+		explicit TexturePack(uint32_t size = 1024U);
 
 		TexturePack(const TexturePack&) = delete;
 		TexturePack(TexturePack&&) = delete;
@@ -40,8 +40,8 @@ namespace rawrbox {
 
 		[[nodiscard]] size_t getSpriteCount() const;
 
-		bool canInsertNode(uint16_t width, uint16_t height);
-		rawrbox::PackNode& addSprite(uint16_t width, uint16_t height, const std::vector<uint8_t>& data);
+		bool canInsertNode(uint32_t width, uint32_t height);
+		rawrbox::PackNode& addSprite(uint32_t width, uint32_t height, const std::vector<uint8_t>& data);
 
 		void upload(Diligent::TEXTURE_FORMAT format = Diligent::TEXTURE_FORMAT::TEX_FORMAT_UNKNOWN, bool dynamic = false) override;
